Add trace fields to sample_hook
This is useful in order to sample according to fields that are on the
trace, and not on the span. This way, we can make sure that all traces
are full, i.e. we do not send spans without sending its parents and we
do not send traces without its children.